requirements.txt
# Core Data Processing
pandas==2.1.0
numpy==1.24.0
pandas-ta==0.3.14b
yfinance==0.2.28
ccxt==4.0.0
alpha-vantage==2.3.1

# Machine Learning & AI
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.0.0
tensorflow==2.13.0
transformers==4.30.0
anthropic==0.28.0
langchain==0.1.0

# Reinforcement Learning
gymnasium==0.29.0
stable-baselines3==2.1.0
ray[rllib]==2.7.0

# Portfolio Optimization
pypfopt==1.5.4
cvxpy==1.3.0
scipy==1.11.0

# Technical Analysis
TA-Lib==0.4.27

# News & Sentiment
newsapi-python==0.2.7
beautifulsoup4==4.12.0
feedparser==6.0.10
vaderSentiment==3.3.2

# Trading & Brokers
alpaca-trade-api==3.0.0
ib-insync==0.9.86
python-binance==1.0.19

# Backtesting
backtrader==1.9.78.123
vectorbt==0.25.0

# Visualization
streamlit==1.28.0
plotly==5.17.0
matplotlib==3.7.0
seaborn==0.12.0
bokeh==3.2.0

# Time Series
statsmodels==0.14.0
prophet==1.1.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
sqlalchemy==2.0.0
redis==5.0.0
celery==5.3.0

# Monitoring & Logging
loguru==0.7.0
prometheus-client==0.17.0
sentry-sdk==1.32.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
hypothesis==6.87.0
